referencia/requirements.txt
requests
pandas
matplotlib
python-docx
fpdf2
openpyxl
py-telegram-bot-api
python-dotenv
pyTelegramBotAPI